Allowing to ignore some protocols when scheduling
When scheduling a project, it could be useful to skip some of them in order to wait until the user gets ready to manually launch it.

For example, in a streaming workflow where some parameter remains unknown at the beginning but after the first steps of the processing it becomes known. Then, the protocol that needs this parameter can be updated and launched without any disruption, letting continue with the rest of the workflow.